refactor(core): centralize common contract function in core/deprecatedstate pkgs
